update new test state with colors according to wrong or right attempt
there is a test there but it will be removed, as well as almost this
entire implementation since this is a prototype and im just playing
around to see what i even seen to make this work. After i figure it out
I will be nuking this and starting over.